Adding drivetrainTicks to SpikeMarks

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Robots/Octonaut.java
@@ -33,7 +33,7 @@ public class Octonaut extends RobotBase {
     public Octonaut(@NonNull HardwareMap hardwareMap,
         LinearOpMode opModeInstance,
         int drivetrainTicks,
-        int wheelDiameter,
+        double wheelDiameter,
         int odoTicks,
         float odoWheelDiam,
         float odoWidth,
